assets_generator/requirements.txt
elevenlabs
pydub
python-dotenv
sentence-transformers